Moving common code to shared
