+++ .env
@@ -4,10 +4,8 @@ INPUT_FILENAMES="**\AssemblyInfo.cs,\n**\AssemblyInfo.vb,\n**\GlobalInfo.cs\n**\
 #INPUT_FILENAMES="AssemblyInfo.cs\r\nAssemblyInfo.vb"
 INPUT_INSERTATTRIBUTES=true
 INPUT_FILEENCODING = "auto"
-#INPUT_FILEENCODING = "utf-16le"
 INPUT_WRITEBOM=true
 
-#INPUT_VERSIONNUMBER="TS Extension Test Build_2018.11.*"
 INPUT_PACKAGEVERSION="9.8.7-beta65"
 INPUT_ASSEMBLYVERSIONNUMBER="TS Extension Test Build_2018.11.*"
 INPUT_FILEVERSIONNUMBER="1990.03.*.*"
@@ -23,7 +21,7 @@ INPUT_CULTURE="en-GB"
 # INPUT_COPYRIGHT="Copyright © $(date:yyyy) $(date:dd.MM.yyyy dd MMMM yyyy HH:mm tt) Example Ltd"
 
 # Node (Moment JS)
-INPUT_COPYRIGHT="Copyright © $(date:YYYY) $(date:DD.MM.YYYY DD MMMM YYYY HH:mm a) Example Ltd";
+INPUT_COPYRIGHT="Copyright © $(date:YYYY) $(date:DD.MM.YYYY DD MMMM YYYY HH:mm a) Example Ltd"
 
 # Net Framework
 INPUT_TRADEMARK="Example Trademark"
