+++ package.json
@@ -27,8 +27,8 @@
     "@types/chardet": "^0.5.0",
     "@types/xml2js": "^0.4.3",
     "chardet": "^0.7.0",
+    "iconv-lite": "^0.4.24",
     "moment": "^2.22.2",
-    "read-xml": "^3.0.0",
     "vsts-task-lib": "^2.6.0",
     "xml2js": "^0.4.19"
   },
@@ -36,7 +36,6 @@
     "@types/mocha": "^5.2.5",
     "@types/node": "^10.9.4",
     "@types/q": "^1.5.0",
-    "iconv-lite": "^0.4.24",
     "ts-node": "^6.2.0",
     "typescript": "^2.9.2"
   }
